nebula-backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-groq
langchain-google-genai
huggingface_hub
langchain-huggingface
requests